cohorts/2024/09-serverless/requirements.txt
tensorflow==2.14.0
pillow
